requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
django-social-share==2.3.0
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
Pillow==9.2.0
python-resize-image==1.1.20
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
